# examples/microservices/userservice/internal/users/dtos/v1/requests/change_password_request.go
package requests

// ChangePasswordRequest DTO for changing password
type ChangePasswordRequest struct {
	OldPassword string `json:"oldPassword" validate:"required"`
	NewPassword string `json:"newPassword" validate:"required,min=8"`
}