# examples/microservices/userservice/internal/users/dtos/v1/requests/signin_request.go
package requests

// SignInRequest DTO for user authentication
type SignInRequest struct {
	Username string `json:"username" validate:"required"`
	Password string `json:"password" validate:"required"`
}